% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootiter}
\alias{bootiter}
\title{bootiter}
\usage{
bootiter(i, qteparams, func)
}
\arguments{
\item{i}{the bootstrap iteration, this is not used except for calling
by apply-type methods}

\item{qteparams}{QTEparams object containing the parameters passed to ciqte}

\item{func}{the internal function that computes estimates of the qte,
should be the same as what was used to produce qteest}
}
\value{
SE object
}
\description{
\code{bootiter} that computes a single bootstrap iteration
}
\keyword{internal}
