% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(qteparams, qteest, func)
}
\arguments{
\item{qteparams}{QTEparams object containing the parameters passed to ciqte}

\item{qteest}{an initial estimate of the qte}

\item{func}{the internal function that computes estimates of the qte,
should be the same as what was used to produce qteest}
}
\value{
SE object
}
\description{
\code{bootstrap} is a method that bootstraps standard
errors for methods in the qte package
}
\keyword{internal}
