% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{BoundsObj}
\alias{BoundsObj}
\title{BoundsObj}
\usage{
BoundsObj(lbs, ubs, ub.quantiles, lb.quantiles, ub.qte, lb.qte,
  att = NULL, probs)
}
\arguments{
\item{lbs}{A vector of the lower bounds for each value in the support
of the outcome}

\item{ubs}{A vector of the upper bounds for each value in the support
of the outcome}

\item{ub.quantiles}{A vector of the same length as probs that contains
the upper bound of the quantiles of the counterfactual distribution
of untreated potential outcomes for the treated group}

\item{lb.quantiles}{A vector of the same length as probs that contains
the lower bound of the quantiles of the counterfactual distribution
of untreated potential outcomes for the treated group}

\item{ub.qte}{The point estimate of the upper bound for the QTE}

\item{lb.qte}{The point estimate of the lower bound for the QTE}

\item{att}{The ATT is point identified under the assumptions required
by the bounds method}

\item{probs}{A vector of values between 0 and 1 to compute the QTET at}
}
\description{
An object of results from computing bounds
}
\keyword{internal}
