% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qte.R
\name{compute.ci.qte}
\alias{compute.ci.qte}
\title{compute.ci.qte}
\usage{
compute.ci.qte(formla, x = NULL, data, probs = seq(0.05, 0.95, 0.05),
  method = "logit")
}
\arguments{
\item{formla}{The formula y ~ d where y is the outcome and d is the
treatment indicator (d should be binary)}

\item{x}{Vector of covariates.  Default is no covariates}

\item{data}{The name of the data.frame that contains the data}

\item{probs}{A vector of values between 0 and 1 to compute the QTET at}

\item{method}{Method to compute propensity score.  Default is logit; other
 option is probit.}
}
\value{
QTE object
}
\description{
compute.ci.qte
}
\keyword{internal}

