% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiC.R
\name{compute.CiC}
\alias{compute.CiC}
\title{compute.CiC}
\usage{
compute.CiC(qp)
}
\value{
QTE object
}
\description{
\code{compute.CiC} does the computational work for the Change in Changes
model of Athey and Imbens (2006). Called both for the point estimate and
inside each bootstrap iteration.
}
\keyword{internal}
