% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\name{compute.panel.qtet}
\alias{compute.panel.qtet}
\title{compute.panel.qtet}
\usage{
compute.panel.qtet(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to ciqte}
}
\value{
QTE object
}
\description{
\code{compute.panel.qtet} uses third period of data,
combined with Distributional
Difference in Differences assumption (Fan and Yu, 2012)
to point identify QTET.
}
