% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QDiD.R
\name{compute.QDiD}
\alias{compute.QDiD}
\title{Quantile Difference in Differences}
\usage{
compute.QDiD(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to QDiD}
}
\value{
QTE object
}
\description{
\code{compute.QDiD} computes the Quantile Difference in
Differences estimator
}
\keyword{internal}
