% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatt.R
\name{compute.spatt}
\alias{compute.spatt}
\title{compute.spatt}
\usage{
compute.spatt(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to spatt}
}
\value{
QTE object with \code{ate} set to the ATT estimate and \code{qte}
set to \code{NULL} (no quantile estimates).
}
\description{
\code{compute.spatt} implements the Abadie (2005) semiparametric DiD
estimator for the ATT. This is called internally by \code{spatt} but can
also be used directly when quantile treatment effects are not needed.
}
\keyword{internal}
