% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_qte.R
\name{compute.unc_qte}
\alias{compute.unc_qte}
\title{compute.unc_qte}
\usage{
compute.unc_qte(qp)
}
\arguments{
\item{qp}{QTEparams object}
}
\value{
QTE object
}
\description{
Computes the QTE under unconfoundedness using Firpo (2007)
propensity score reweighting. Called internally by \code{unc_qte} and
during each bootstrap iteration.
}
\keyword{internal}
