% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\name{computeDiffSE}
\alias{computeDiffSE}
\title{computeDiffSE}
\usage{
computeDiffSE(est1, bootIters1, est2, bootIters2, alp = 0.05)
}
\arguments{
\item{est1}{A QTE object containing the first set of estimates}

\item{bootIters1}{A List of QTE objects that have been bootstrapped}

\item{est2}{A QTE object containing a second set of estimates}

\item{bootIters2}{A List of QTE objects that have been bootstrapped
using the second method}

\item{alp}{The significance level used for constructing bootstrap
confidence intervals}
}
\description{
Takes two sets of initial estimates and bootstrap
estimations
(they need to have the same number of iterations) and determines
whether or not the estimates are statistically different from each
other.  It can be used to compare any sets of estimates, but it is
particularly used here to compare estimates from observational methods
with observations from the experimental data (which also have standard
errors because, even though the estimates are cleanly identified, they
are still estimated).
}
