% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cs2panel}
\alias{cs2panel}
\title{cs2panel}
\usage{
cs2panel(cs1, cs2, yname)
}
\arguments{
\item{cs1}{data frame, the first cross section}

\item{cs2}{data frame, the second cross section}

\item{yname}{the name of the variable to calculate difference for (should be the same in each dataset)}
}
\value{
the change in outcomes over time
}
\description{
Turn repeated cross sections data into panel data by imposing rank invariance; does not
 that the inputs have the same length
}
