% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{diffQ}
\alias{diffQ}
\title{diffQ}
\usage{
diffQ(qvec, tauvec, hightau, lowtau)
}
\arguments{
\item{qvec}{vector of quantiles}

\item{tauvec}{vector of tau (should be same length as qvec)}

\item{hightau}{upper quantile}

\item{lowtau}{lower quantile}
}
\value{
scalar difference between quantiles
}
\description{
\subsection{takes a single set of quantiles}{

(not qtes as in diffquantiles)
and returns the difference between particular ones
}
}
