% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{dr}
\alias{dr}
\title{dr}
\usage{
dr(formla, data, y.seq)
}
\arguments{
\item{formla}{the regression to run}

\item{y.seq}{the values of y to run the regression on}
}
\description{
Distribution Regression
}
