% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR.R
\name{DR}
\alias{DR}
\title{DR}
\usage{
DR(yvals, drlist)
}
\arguments{
\item{yvals}{A vector of values that y can take}

\item{drlist}{A list where for each value of y, a distribution regression}
}
\description{
A distribution regression object
}
