% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{getWeightedMean}
\alias{getWeightedMean}
\title{getWeightedMean}
\usage{
getWeightedMean(y, weights = NULL, norm = TRUE)
}
\arguments{
\item{y}{a vector to compute the mean for}

\item{weights}{the vector of weights, can be NULL, then will just return mean}

\item{norm}{normalize the weights so that they have mean of 1, default is
to normalize}
}
\description{
Get the mean applying some weights
}
\keyword{internal}
