% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{getWeightedQuantile}
\alias{getWeightedQuantile}
\title{getWeightedQuantile}
\usage{
getWeightedQuantile(tau, cvec, weights = NULL, norm = TRUE)
}
\arguments{
\item{tau}{between 0 and 1, ex. .5 implies get the median}

\item{cvec}{a vector to compute quantiles for}

\item{weights}{the weights, weighted.checkfun normalizes the weights
to sum to 1.}

\item{norm}{normalize the weights so that they have mean of 1, default is
to normalize}
}
\description{
Finds the quantile by optimizing the weighted check function
}
\keyword{internal}
