% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{ggplot.qte}
\alias{ggplot.qte}
\title{ggplot.qte}
\usage{
\method{ggplot}{qte}(qteobj)
}
\arguments{
\item{qteobj}{a QTE object}
}
\value{
a ggplot object
}
\description{
Makes somewhat nicer plots of quantile treatment effects
 by using ggplot
}
