% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{ids2rownum}
\alias{ids2rownum}
\title{ids2rownum}
\usage{
ids2rownum(ids, data, idname)
}
\arguments{
\item{ids}{vector of ids}

\item{data}{data frame}

\item{idname}{unique id}
}
\description{
ids2rownum takes a vector of ids and converts it t the right
 row number in the dataset; ids should be unique in the dataset
 that is, don't pass the function panel data with multiple same ids
}
\keyword{internal}
