% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\docType{data}
\name{lalonde.exp.panel}
\alias{lalonde.exp.panel}
\title{Lalonde's Panel Experimental Dataset}
\description{
The panel verion of the experimental part of the
 \code{lalonde} dataset.  It
 is loaded with all the datasets with the command \code{data(lalonde)}
}
\keyword{datasets}
