% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\docType{data}
\name{lalonde.exp}
\alias{lalonde.exp}
\title{Lalonde's Experimental Dataset}
\description{
The cross sectional verion of the experimental part of the
\code{lalonde} dataset.  It
is loaded with all the datasets with the command \code{data(lalonde)}
}
\keyword{datasets}
