% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{panel2cs}
\alias{panel2cs}
\title{panel2cs}
\usage{
panel2cs(data, timevars, idname, tname)
}
\arguments{
\item{data}{data.frame used in function}

\item{timevars}{vector of names of variables to keep}

\item{idname}{unique id}

\item{tname}{time period name}
}
\description{
panel2cs takes a 2 period dataset and turns it
 into a cross sectional dataset.  The default functionality
 is to keep all the variables from period 1
 and add all the variables listed by name in timevars
 from period 2 to those
}
\keyword{internal}
