% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-checks.R
\name{panelize.data}
\alias{panelize.data}
\title{panelize.data}
\usage{
panelize.data(data, idname, tname, t, tmin1, tmin2 = NULL)
}
\arguments{
\item{data}{A data.frame containing all the variables used}

\item{idname}{The individual (cross-sectional unit) id name}

\item{tname}{The name of the column containing the time periods}

\item{t}{The 3rd time period in the sample.  Treated individuals should
be treated in this time period and untreated individuals should not be
treated.  The code attempts to enforce this condition, but it is good
try to handle this outside the panel.qtet method.}

\item{tmin1}{The 2nd time period in the sample.  This should be a
pre-treatment period for all individuals in the sample.}

\item{tmin2}{The 1st time period in the sample.  This should be a
pre-treatment period for all individuals in the sample.}
}
\value{
data.frame
}
\description{
get data in correct format for using panel methods in qte package
}
