% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{plot.BoundsObj}
\alias{plot.BoundsObj}
\title{Plot Bounds}
\usage{
\method{plot}{BoundsObj}(x, plotate = FALSE, plot0 = FALSE,
  qtecol = "black", atecol = "black", col0 = "black", ylim = NULL,
  uselegend = FALSE, legloc = "topright", ...)
}
\arguments{
\item{x}{A BoundsObj Object}

\item{plotate}{Boolean whether or not to plot the ATE}

\item{plot0}{Boolean whether to plot a line at 0}

\item{qtecol}{Color for qte plot.  Default "black"}

\item{atecol}{Color for ate plot.  Default "black"}

\item{col0}{Color for 0 plot.  Default "black"}

\item{ylim}{The ylim for the plot; if not passed, it will be automatically
set based on the values that the QTE takes}

\item{uselegend}{Boolean whether or not to print a legend}

\item{legloc}{String location for the legend.  Default "topright"}

\item{...}{Other parameters to be passed to plot (e.g lwd)}
}
\description{
Plots a BoundObj Object
}
