% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{print.matrix2}
\alias{print.matrix2}
\title{print.matrix2}
\usage{
\method{print}{matrix2}(m, header = NULL, digits = 2, nsmall = 2)
}
\arguments{
\item{m}{Some matrix}
}
\description{
Helper function to print a matrix; used by the print methods
}
\keyword{internal}
