% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte-class.R
\name{print.summary.QTE}
\alias{print.summary.QTE}
\title{Print summary.QTE}
\usage{
\method{print}{summary.QTE}(x, ...)
}
\arguments{
\item{x}{A \code{summary.QTE} object.}

\item{...}{unused.}
}
\value{
None. Called for its side effect of printing.
}
\description{
Prints a \code{summary.QTE} object.
}
