% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{qte-package}
\alias{qte}
\alias{qte-package}
\title{qte: A package for computing quantile treatment effects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides several methods for computing the Quantile Treatment Effect (QTE) and Quantile Treatment Effect on the Treated (QTT). The main cases covered are (i) Treatment is randomly assigned, (ii) Treatment is as good as randomly assigned after conditioning on some covariates (also called conditional independence or selection on observables) using the methods developed in Firpo (2007) \doi{10.1111/j.1468-0262.2007.00738.x}, (iii) Identification is based on a Difference in Differences assumption (several varieties are available in the package e.g. Athey and Imbens (2006) \doi{10.1111/j.1468-0262.2006.00668.x} Callaway and Li (2019) \doi{10.3982/QE935}, Callaway, Li, and Oka (2018) \doi{10.1016/j.jeconom.2018.06.008}).
}
\author{
\strong{Maintainer}: Brantly Callaway \email{brantly.callaway@uga.edu}

Authors:
\itemize{
  \item Brantly Callaway \email{brantly.callaway@uga.edu}
}

}
