% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{qtes2mat}
\alias{qtes2mat}
\title{qtes2mat}
\usage{
qtes2mat(qteList, sset = NULL, se = TRUE, rnd = 3)
}
\arguments{
\item{qteList}{a list of qte objects}

\item{sset}{subset of qtes to keep}

\item{se}{whether or not to include standard errors in the resulting matrix}

\item{rnd}{how many disgits to round to}
}
\value{
matrix
}
\description{
Turn multiple qtes into a matrix for printing
}
