% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{qteToTexreg}
\alias{qteToTexreg}
\title{diffQ}
\usage{
qteToTexreg(qteobj, tau = NULL, reportAte = T)
}
\arguments{
\item{qteobj}{A QTE object}

\item{tau}{Optional vector of taus to texreg results for}

\item{reportAte}{Whether or not texreg the ATE (or ATT) as well}
}
\description{
\subsection{takes a single set of quantiles}{

(not qtes as in diffquantiles)
and returns the difference between particular ones
}
}
