% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\name{SE}
\alias{SE}
\title{SE}
\usage{
SE(qte.se = NULL, ate.se = NULL, qte.upper = NULL,
  qte.lower = NULL, ate.upper = NULL, ate.lower = NULL, c = NULL,
  probs = NULL)
}
\arguments{
\item{qte.se}{The QTE Standard Error}

\item{ate.se}{The ATE Standard Error}

\item{qte.upper}{The QTE upper CI}

\item{qte.lower}{The QTE lower CI}

\item{ate.upper}{The ATE upper CI}

\item{ate.lower}{The ATE lower CI}

\item{c}{The critical value from a KS-type statistic used for creating
uniform confidence bands}

\item{probs}{The values at which the QTE is computed}
}
\description{
Class for Standard Error Objects
}
\keyword{internal}
