% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qte.R
\name{summary.BoundsObj}
\alias{summary.BoundsObj}
\title{Summary of BoundsObj}
\usage{
\method{summary}{BoundsObj}(object, ...)
}
\arguments{
\item{object}{A BoundsObj Object}

\item{...}{Other params (for consistency as generic S3 method, but not used)}
}
\value{
summary.BoundsObj Object
}
\description{
Summary of BoundsObj
}

