% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiC.R
\name{cic2}
\alias{cic2}
\title{cic2}
\usage{
cic2(
  yname,
  gname,
  tname,
  idname,
  data,
  xformla = ~1,
  ret_quantile = NULL,
  gt_type = "att",
  anticipation = 0,
  cband = TRUE,
  alp = 0.05,
  boot_type = "empirical",
  biters = 100,
  cl = 1
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{ret_quantile}{This parameter determines which quantile will be reported
by the cic2 function.  By default \code{ret_quantile=NULL}; in this case, the
function will return an estimate of the overall ATT and an event study for
the ATT.  Other choices should be between 0 and 1.  For example, if the
user specifies \code{ret_quantile=0.9}, then the function will return overall
and event study parameters for the QTT(0.9).  These ...would be better to return the overall distribution and then to average and invert in later steps...}

\item{alp}{significance level; default is 0.05}

\item{boot_type}{should be one of "multiplier" (the default) or "empirical".
The multiplier bootstrap is generally much faster, but \code{attgt_fun} needs
to provide an expression for the influence function (which could be challenging
to figure out).  If no influence function is provided, then the \code{pte}
package will use the empirical bootstrap no matter what the value of this
parameter.}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}

\item{ret_dist}{If set to be true, the function returns the observed
distribution of outcomes and counterfactual distribution of outcomes
for each (g,t) through the \code{extra_gt_returns} element of \code{group_time_att}
object.}
}
\description{
This is a multi-period implementation of the change-in-changes
approach from Athey and Imbens (2006, Econometrica).  This function
is in a beta release and users should use caution when using this function
in emprical work.

The function builds on the \code{pte} package and will return an overall
treatment effect parameter as well as an event study.  See, in particular,
the argument \code{ret_quantile} below.
}
