% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiC.R
\name{cic_attgt}
\alias{cic_attgt}
\title{cic_attgt}
\usage{
cic_attgt(gt_data, xformla = ~1, ...)
}
\arguments{
\item{gt_data}{data that is "local" to a particular group-time average
treatment effect}

\item{xformla}{one-sided formula for covariates used in the propensity score
and outcome regression models}

\item{...}{extra function arguments; not used here}
}
\value{
pte::attgt_noif object.  \code{cic_attgt} computes attgt using
the CIC approach.  It also returns distributions of observed outcomes
for the treated group (F1), the counterfactual distribution
of untreated potential potential outcomes for the treated group (F0),
and the distribution of the treatment effect under the assumption
of rank invariance over time (Fte) all through the extra_gt_returns
argument to \code{pte::attgt_noif} object.
}
\description{
cic_attgt
}
