% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciqte.R
\name{compute.ci.qte}
\alias{compute.ci.qte}
\title{compute.ci.qte}
\usage{
compute.ci.qte(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to ciqte}
}
\value{
QTE object
}
\description{
\code{compute.ci.qte} computes the QTE under selection
on observables
}
\keyword{internal}
