% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciqtet.R
\name{compute.ci.qtet}
\alias{compute.ci.qtet}
\title{compute.ci.qtet}
\usage{
compute.ci.qtet(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to ciqte}
}
\value{
QTE object
}
\description{
\code{compute.ci.qtet} computes the QTT under
selection on observables
}
\keyword{internal}
