% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiC.R
\name{compute.CiC}
\alias{compute.CiC}
\title{athey.imbens}
\usage{
compute.CiC(qp)
}
\description{
\code{compute.CiC} does the computational
work for the Change in Changes model
of Athey and Imbens, 2006.
}
\keyword{internal}
