% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qte.R
\name{compute.CiC}
\alias{compute.CiC}
\title{athey.imbens}
\usage{
compute.CiC(formla, t, tmin1, tname, x = NULL, data,
  dropalwaystreated = TRUE, panel = FALSE, plot = FALSE, idname = NULL,
  uniqueid = NULL, probs = seq(0.05, 0.95, 0.05))
}
\arguments{
\item{formla}{The formula y ~ d where y is the outcome and d is the
treatment indicator (d should be binary)}

\item{t}{The 3rd time period in the sample (this is the name of the column)}

\item{tmin1}{The 2nd time period in the sample (this is the name of the
column)}

\item{tname}{The name of the column containing the time periods}

\item{x}{A vector of covariates (the name of the columns)}

\item{data}{The name of the data.frame that contains the data}

\item{dropalwaystreated}{How to handle always treated observations
in panel data case (not currently used)}

\item{plot}{Boolean whether or not the estimated QTET should be plotted}

\item{idname}{The individual (cross-sectional unit) id name}

\item{uniqueid}{Not sure what this does}

\item{probs}{A vector of values between 0 and 1 to compute the QTET at}
}
\description{
athey.imbens
}
\keyword{internal}

