% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddid2.R
\name{compute.ddid2}
\alias{compute.ddid2}
\title{compute.ddid2}
\usage{
compute.ddid2(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to ddid2}
}
\value{
QTE object
}
\description{
\code{compute.ddid2} uses two periods of data (repeated cross sections
or panel) to estimate the Quantile Treatment Effect on the Treated (QTET)
}
\keyword{internal}
