% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\name{compute.spatt}
\alias{compute.spatt}
\title{compute.spatt}
\usage{
compute.spatt(formla, xformla = NULL, t, tmin1, tname, x = NULL, data,
  weights = NULL, panel = FALSE, dropalwaystreated = TRUE, idname,
  bootstrap.iter = FALSE, plot = FALSE, method = c("logit", "GMM",
  "semiparametric", "simulation", "direct"), ydiscrete = FALSE)
}
\arguments{
\item{formla}{outcome variable on treatment}

\item{t}{last time period}

\item{tmin1}{middle time period}

\item{x}{additional covariates if using propensity score
reweighting technique}

\item{dropalwaystreated}{boolean indicating whether in true panel data
context (when treatment can occur in any period) whether or
not previously treated observations should be dropped from the sample.
This functionality is not currently implemented}

\item{idname}{an id that identifies individual units over time}

\item{bootstrap.iter}{boolean passed that is passed in when this
method is used to compute standard errors}

\item{plot}{Binary indicating whether or not to plot the object}

\item{method}{How to estimate the propensity score}

\item{ydiscrete}{Used to indicate whether the outcome has any
discrete parts in the distribution}
}
\value{
QTE object
}
\description{
\code{compute.spatt} implements the method of Abadie (2005); this is
 computed automatically in several other methods in the qte package
 but this function provides a standalone result when quantiles are not
 wanted/needed.
}
\keyword{internal}

