% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatt.R
\name{compute.spatt}
\alias{compute.spatt}
\title{compute.spatt}
\usage{
compute.spatt(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to ciqte}
}
\value{
QTE object
}
\description{
\code{compute.spatt} implements the method of Abadie (2005); this is
 computed automatically in several other methods in the qte package
 but this function provides a standalone result when quantiles are not
 wanted/needed.

\code{compute.ddid2} uses two periods of data (repeated cross sections
or panel) to estimate the Quantile Treatment Effect on the Treated (QTET)
}
\keyword{internal}
