% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{getWeightedQuantiles}
\alias{getWeightedQuantiles}
\title{getWeightedQuantile}
\usage{
getWeightedQuantiles(tau, cvec, weights)
}
\arguments{
\item{tau}{a vector of values between 0 and 1}

\item{cvec}{a vector to compute quantiles for}

\item{weights}{the weights, weighted.checkfun normalizes the weights
to sum to 1.}
}
\description{
Finds multiple quantiles by repeatedly calling
 getWeightedQuantile
}

