% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{ggqte}
\alias{ggqte}
\title{ggqte}
\usage{
ggqte(qteobj, main = "", ylab = "", ylim = NULL, ybreaks = NULL,
  xbreaks = c(0.1, 0.3, 0.5, 0.7, 0.9))
}
\arguments{
\item{qteobj}{a QTE object}

\item{main}{optional title}

\item{ylab}{optional y axis label}

\item{ylim}{optional limits of y axis}

\item{ybreaks}{optional breaks in y axis}

\item{xbreaks}{optional breaks in x axis}
}
\value{
a ggplot object
}
\description{
Makes somewhat nicer plots of quantile treatment effects
 by using ggplot
}
