% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-checks.R
\name{panel.checks}
\alias{panel.checks}
\title{panel.checks}
\usage{
panel.checks(qp)
}
\arguments{
\item{qp}{QTEparams object containing the parameters passed to ciqte}
}
\description{
Does some checking that data setup is valid for using methods
 in qte package
}
