% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qte.R
\name{print.matrix1}
\alias{print.matrix1}
\title{print.matrix1}
\usage{
\method{print}{matrix1}(m)
}
\arguments{
\item{m}{Some matrix}
}
\description{
Helper function to print a matrix; used by the print methods
}
\keyword{internal}

