% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qte.R
\name{print.summary.BoundsObj}
\alias{print.summary.BoundsObj}
\title{Print a summary.BoundsObj}
\usage{
\method{print}{summary.BoundsObj}(x, ...)
}
\arguments{
\item{x}{A summary.BoundsObj}

\item{...}{Other objects to pass (not used)}
}
\description{
Prints a Summary QTE Object
}

