% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\name{print.summary.QTE}
\alias{print.summary.QTE}
\title{Print}
\usage{
\method{print}{summary.QTE}(x, ...)
}
\arguments{
\item{x}{A summary.QTE object}

\item{...}{Other params (required as generic function, but not used)}
}
\description{
Prints a Summary QTE Object
}
