% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{qte-package}
\alias{qte}
\alias{qte-package}
\title{qte: A package for computing quantile treatment effects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides several methods for computing the Quantile Treatment Effect (QTE) and Quantile Treatment Effect on the Treated (QTT). The main cases covered are (i) treatment is randomly assigned, (ii) treatment is as good as randomly assigned after conditioning on covariates (selection on observables) using the methods of Firpo (2007) \doi{10.1111/j.1468-0262.2007.00738.x}, and (iii) identification is based on a Difference in Differences assumption, with support for several varieties including Athey and Imbens (2006) \doi{10.1111/j.1468-0262.2006.00668.x}, Callaway and Li (2019) \doi{10.3982/QE935}, and Callaway, Li, and Oka (2018) \doi{10.1016/j.jeconom.2018.06.008}. Version 2.0 adds a unified staggered treatment adoption API (built on 'ptetools') for all DiD-based estimators, as well as a new lagged-outcome unconfoundedness estimator ('lou_qte').
}
\author{
\strong{Maintainer}: Brantly Callaway \email{brantly.callaway@uga.edu}

Authors:
\itemize{
  \item Brantly Callaway \email{brantly.callaway@uga.edu}
}

}
