% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{setupData}
\usage{
setupData(qteParams)
}
\arguments{
\item{qteParams}{object holding the function parameters}
}
\value{
qteData object holding data to be used in QTE functions
}
\description{
\code{setupData} sets up the data to use in each
 compute method in the QTE package
}
