% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{summary.BoundsObj}
\alias{summary.BoundsObj}
\title{Summary of BoundsObj}
\usage{
\method{summary}{BoundsObj}(object, ...)
}
\arguments{
\item{object}{A BoundsObj Object}

\item{...}{Other params (for consistency as generic S3 method, but not used)}
}
\value{
summary.BoundsObj Object
}
\description{
\code{summary.BoundsObj} is an object for holding
\code{bounds} results
}
