% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qte.R
\name{summary.QTE}
\alias{summary.QTE}
\title{Summary}
\usage{
\method{summary}{QTE}(object, ...)
}
\arguments{
\item{object}{A QTE Object}

\item{...}{Other params (to work as generic method, but not used)}
}
\description{
\code{summary.QTE} summarizes QTE objects
}
